requirements.txt
streamlit
pdfplumber
pandas
tabula-py
openpyxl